requirements.txt
jinja2
tqdm
absl-py
openai
pyyaml
google-genai
anthropic
marko
pandas
fastapi
uvicorn[standard]
pydantic
livekit
livekit-agents
livekit-plugins-openai
aiohttp